feat: tidy, merged collectInputFiles for each ast type
